refactor: move `Jdg` to syntax so we no longer need `TyckResult`
