+++ base/src/main/java/org/aya/tyck/ExprTycker.java
@@ -11,6 +11,7 @@ import org.aya.generic.term.DTKind;
 import org.aya.pretty.doc.Doc;
 import org.aya.syntax.concrete.Expr;
 import org.aya.syntax.core.Closure;
+import org.aya.syntax.core.Jdg;
 import org.aya.syntax.core.def.DataDefLike;
 import org.aya.syntax.core.def.PrimDef;
 import org.aya.syntax.core.repr.AyaShape;
