+++ base/src/main/java/org/aya/tyck/pat/ClauseTycker.java
@@ -11,6 +11,7 @@ import org.aya.normalize.Finalizer;
 import org.aya.prettier.AyaPrettierOptions;
 import org.aya.syntax.concrete.Expr;
 import org.aya.syntax.concrete.Pattern;
+import org.aya.syntax.core.Jdg;
 import org.aya.syntax.core.pat.Pat;
 import org.aya.syntax.core.pat.TypeEraser;
 import org.aya.syntax.core.term.ErrorTerm;
@@ -21,7 +22,6 @@ import org.aya.syntax.ref.LocalCtx;
 import org.aya.syntax.ref.LocalVar;
 import org.aya.syntax.telescope.Signature;
 import org.aya.tyck.ExprTycker;
-import org.aya.tyck.Jdg;
 import org.aya.tyck.TyckState;
 import org.aya.tyck.ctx.LocalLet;
 import org.aya.tyck.error.PatternProblem;
