+++ base/src/main/java/org/aya/tyck/tycker/AppTycker.java
@@ -12,6 +12,7 @@ import org.aya.syntax.compile.JitData;
 import org.aya.syntax.compile.JitFn;
 import org.aya.syntax.compile.JitPrim;
 import org.aya.syntax.concrete.stmt.decl.*;
+import org.aya.syntax.core.Jdg;
 import org.aya.syntax.core.def.*;
 import org.aya.syntax.core.repr.AyaShape;
 import org.aya.syntax.core.term.Term;
@@ -19,7 +20,6 @@ import org.aya.syntax.core.term.call.*;
 import org.aya.syntax.ref.DefVar;
 import org.aya.syntax.ref.LocalVar;
 import org.aya.syntax.telescope.AbstractTele;
-import org.aya.tyck.Jdg;
 import org.aya.tyck.TyckState;
 import org.aya.util.error.Panic;
 import org.aya.util.error.SourcePos;
