+++ base/src/main/java/org/aya/tyck/tycker/TeleTycker.java
@@ -8,6 +8,7 @@ import kala.collection.mutable.MutableList;
 import kala.collection.mutable.MutableSeq;
 import org.aya.normalize.Finalizer;
 import org.aya.syntax.concrete.Expr;
+import org.aya.syntax.core.Jdg;
 import org.aya.syntax.core.term.ErrorTerm;
 import org.aya.syntax.core.term.Param;
 import org.aya.syntax.core.term.SortTerm;
@@ -17,7 +18,6 @@ import org.aya.syntax.ref.LocalVar;
 import org.aya.syntax.telescope.AbstractTele;
 import org.aya.syntax.telescope.Signature;
 import org.aya.tyck.ExprTycker;
-import org.aya.tyck.Jdg;
 import org.aya.tyck.error.UnifyError;
 import org.aya.unify.Synthesizer;
 import org.aya.util.error.SourcePos;
