+++ cli-impl/src/main/java/org/aya/cli/interactive/ReplCompiler.java
@@ -33,6 +33,7 @@ import org.aya.resolve.visitor.ExprResolver;
 import org.aya.syntax.GenericAyaFile;
 import org.aya.syntax.concrete.Expr;
 import org.aya.syntax.concrete.stmt.Stmt;
+import org.aya.syntax.core.Jdg;
 import org.aya.syntax.core.def.PrimDef;
 import org.aya.syntax.core.def.TyckDef;
 import org.aya.syntax.core.term.Term;
@@ -40,7 +41,6 @@ import org.aya.syntax.literate.CodeOptions.NormalizeMode;
 import org.aya.syntax.ref.AnyVar;
 import org.aya.syntax.ref.ModulePath;
 import org.aya.tyck.ExprTycker;
-import org.aya.tyck.Jdg;
 import org.aya.tyck.TyckState;
 import org.aya.tyck.tycker.TeleTycker;
 import org.aya.util.error.SourceFileLocator;
