+++ syntax/src/main/java/org/aya/syntax/core/Jdg.java
@@ -1,6 +1,6 @@
 // Copyright (c) 2020-2024 Tesla (Yinsen) Zhang.
 // Use of this source code is governed by the MIT license that can be found in the LICENSE.md file.
-package org.aya.tyck;
+package org.aya.syntax.core;
 
 import kala.value.LazyValue;
 import org.aya.syntax.core.term.SortTerm;
